requirements.txt
openai==1.39.0
PyPDF2
python-dotenv
